feat: moving fix to match sqlfluff
