+++ src/core/parser/segments.rs
@@ -1,4 +1,5 @@
 pub mod base;
 pub mod file;
+pub mod fix;
 pub mod meta;
 pub mod raw;
